requirements.txt
streamlit>=1.25
pandas>=2.0
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
plotly.express
plotly.graph_objects
geopy
tiktoken
google-generativeai
